Revise log Gaussian Cox process
A significant revision to continuous and discrete-time log Gaussian Cox processes.

Both type definitons are simplified by requiring all node to use the same grid points and bias value.

Discrete-time models are essentially equivalent to continuous-time models, but store lambdas in a matrix to match discrete-time data format and require a time step parameter.

All methods are updated to work with the new type definitions, and all discrete-time examples are running successfully.

Update includes additional minor improvements to syntax, etc.

+++ src/utils/interpolation.jl
@@ -76,7 +76,7 @@ A discrete linear interpolater between values `y` evaluated at gridpoints `x`.
 - `y::Vector{Float64}`: a vector of function values
 """
 struct DiscreteLinearInterpolator <: Interpolator
-    x::Vector{Int64}
+    x::Vector{Float64}
     y::Vector{Float64}
     I::Union{Float64,Missing}
 end
